requirements.txt
albumentations==1.4.18
numpy==1.24.4
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
